bd-1716: Tillate desimaltall i input (#33)

+++ src/main/kotlin/no/nav/bidrag/inntekt/util/InntektUtil.kt
@@ -67,10 +67,10 @@ fun String.isNumeric(): Boolean {
     return this.all { it.isDigit() }
 }
 
-fun beregneBeløpPerMåned(beløp: Int, antallMnd: Int): BigDecimal {
+fun beregneBeløpPerMåned(beløp: BigDecimal, antallMnd: Int): BigDecimal {
     return if (antallMnd == 0) {
         BigDecimal.valueOf(0)
     } else {
-        beløp.toBigDecimal().div(antallMnd.toBigDecimal())
+        beløp.div(antallMnd.toBigDecimal())
     }
 }
